firmware: Fix more UBSAN left shift errors.
Follow up commit to CL:1867970.

Shifting a uint8_t left by 24 promotes to an int, not an unsigned int
(and shifts into the sign bit are undefined). Probably doesn't make a
difference in assembly but still doesn't hurt to fix. Courtesy of UBSAN.

BRANCH=None
BUG=chromium:1015908
TEST=No more shift errors in unit test

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/1881475
Tested-by: Manoj Gupta <manojgupta@chromium.org>

+++ firmware/2lib/2rsa.c
@@ -145,7 +145,8 @@ static void modpow(const struct vb2_public_key *key, uint8_t *inout,
 	/* Convert from big endian byte array to little endian word array. */
 	for (i = 0; i < (int)key->arrsize; ++i) {
 		uint32_t tmp =
-			(inout[((key->arrsize - 1 - i) * 4) + 0] << 24) |
+			((uint32_t)inout[((key->arrsize - 1 - i) * 4) + 0]
+				<< 24) |
 			(inout[((key->arrsize - 1 - i) * 4) + 1] << 16) |
 			(inout[((key->arrsize - 1 - i) * 4) + 2] << 8) |
 			(inout[((key->arrsize - 1 - i) * 4) + 3] << 0);
